Remove RetrofitCallMapper class and ChatApi interface
