--- client/src/main/java/io/getstream/chat/android/client/api/RetrofitCallMapper.kt
@@ -1,11 +1,0 @@
-package io.getstream.chat.android.client.api
-
-import io.getstream.chat.android.client.call.Call
-import io.getstream.chat.android.client.call.RetrofitCall
-import io.getstream.chat.android.client.parser.ChatParser
-
-internal class RetrofitCallMapper(private val chatParser: ChatParser) {
-    fun <T : Any> map(call: retrofit2.Call<T>): Call<T> {
-        return RetrofitCall(call, chatParser)
-    }
-}
